Fixed problem with removing the directories

+++ configs.sh
@@ -1,11 +1,34 @@
-rm -R /home/$USER/.config/fastfetch
-rm -R /home/$USER/.config/firefox
-rm -R /home/$USER/.config/hypr
-rm -R /home/$USER/.config/kitty
-rm -R /home/$USER/.config/mako
-rm -R /home/$USER/.config/nvim
-sudo rm -R /home/$USER/.config/waybar
-rm -R /home/$USER/.config/wofi
+if [ -d "/home/$USER/.config/fastfetch"]; then
+  rm -R /home/$USER/.config/fastfetch
+fi
+
+if [ -d "/home/$USER/.config/firefox"]; then
+  rm -R /home/$USER/.config/firefox
+fi
+
+if [-d "/home/$USER/.config/hypr"]; then
+  rm -R /home/$USER/.config/hypr
+fi
+
+if [-d "/home/$USER/.config/kitty"]; then
+  rm -R /home/$USER/.config/kitty
+fi
+
+if [-d "/home/$USER/.config/mako"]; then
+  rm -R /home/$USER/.config/mako
+fi
+
+if [-d "/home/$USER/.config/nvim"]; then
+  sudo rm -R /home/$USER/.config/nvim
+fi
+
+if [-d "/home/$USER/.config/nvim"]; then
+  sudo rm -R /home/$USER/.config/waybar
+fi
+
+if [-d "/home/$USER/.config/wofi"]; then
+  rm -R /home/$USER/.config/wofi
+fi
 
 cp -R dotfiles/hypr /home/$USER/.config/
 cp -R dotfiles/fastfetch /home/$USER/.config/
